+++ Location/FormatFile.m
@@ -1,8 +1,8 @@
 //
 //  FormatFile.m
-//  Location
+//  StudentLife
 //
-//  Created by Student student on 2/3/15.
+//  Created by Aaron Jun Yang on 2/3/15.
 //  Copyright (c) 2015 Location. All rights reserved.
 //
 
@@ -20,9 +20,9 @@
     float minutes = floor((lroundf(value) - hours * 3600)/60); //floor(lroundf(value)/60);
     float seconds = lroundf(value) - (hours * 3600) - (minutes * 60);
     
-    int roundedHours = lroundf(hours);
-    int roundedSeconds = lroundf(seconds);
-    int roundedMinutes = lroundf(minutes);
+    int roundedHours = (int)lroundf(hours);
+    int roundedSeconds = (int)lroundf(seconds);
+    int roundedMinutes = (int)lroundf(minutes);
     
     NSString *time = [[NSString alloc]
                       initWithFormat:@"%02d:%02d:%02d", roundedHours,
