+++ Location/StudentLifeBackgroundAudio.m
@@ -100,9 +100,9 @@
     float minutes = floor((lroundf(value) - hours * 3600)/60); //floor(lroundf(value)/60);
     float seconds = lroundf(value) - (hours * 3600) - (minutes * 60);
     
-    int roundedHours = lroundf(hours);
-    int roundedSeconds = lroundf(seconds);
-    int roundedMinutes = lroundf(minutes);
+    int roundedHours = (int)lroundf(hours);
+    int roundedSeconds = (int)lroundf(seconds);
+    int roundedMinutes = (int)lroundf(minutes);
     
     NSString *time = [[NSString alloc]
                       initWithFormat:@"%02d:%02d:%02d", roundedHours,
